# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: taboo · 禁制 — 密钥管理平台 API
  version: 0.7.0
  description: 开源自部署密钥管理平台。高价值数据 + 严格访问控制 + 全程留痕。M5：Secret Sync / Webhooks / OIDC SSO / 审计导出。
servers:
  - url: http://localhost:7100
    description: 本地开发
...
paths:
...
  /api/v1/orgs/{slug}/sync/targets:
    get:
      tags: [sync]
      summary: 同步目标列表（成员可见；config 只回显已配置的键名）
      operationId: listSyncTargets
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
      responses:
        '200':
          description: 目标列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  targets: { type: array, items: { $ref: '#/components/schemas/SyncTarget' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [sync]
      summary: 创建同步目标（owner；平台凭证 Master Key 加密落库；?ping=1 先自检连通性）
      operationId: createSyncTarget
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SyncTargetInput' }
      responses:
        '201': { description: 已创建, content: { application/json: { schema: { type: object, properties: { id: { type: string }, name: { type: string }, platform: { type: string } } } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/orgs/{slug}/sync/targets/{id}:
    delete:
      tags: [sync]
      summary: 删除同步目标（owner）
      operationId: deleteSyncTarget
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 已删除, content: { application/json: { schema: { type: object, properties: { id: { type: string }, deleted: { type: boolean } } } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/orgs/{slug}/sync/targets/{id}/retry:
    post:
      tags: [sync]
      summary: 手动全量重推（owner；绑定范围内所有密钥入队）
      operationId: retrySyncTarget
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: 已入队, content: { application/json: { schema: { type: object, properties: { target_id: { type: string }, enqueued: { type: integer } } } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/orgs/{slug}/sync/runs:
    get:
      tags: [sync]
      summary: 同步记录（指纹审计，不含明文；可按目标/状态筛选）
      operationId: listSyncRuns
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/OrgSlug' }
        - { name: target_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [pending, success, failed, dead] } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200':
          description: 同步记录
          content:
            application/json:
              schema:
                type: object
                properties:
                  runs: { type: array, items: { $ref: '#/components/schemas/SyncRun' } }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ---------- Webhooks（M5 #12） ----------
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    OrgSlug: { name: slug, in: path, required: true, schema: { type: string } }
...
  responses:
    BadRequest: { description: 参数错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Unauthorized: { description: 未认证/凭证错误, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: 权限不足（RBAC），如 viewer 取明文、developer 写 prod, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    NotFound: { description: 资源不存在, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      properties:
        code: { type: string, example: BAD_CREDENTIALS }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    SyncTarget:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        platform: { type: string, enum: [github, vercel, cloudflare] }
        project_id: { type: string }
        env_slug: { type: string, description: "'*' 为全部环境" }
        enabled: { type: boolean }
        created_at: { type: string }
        config_keys: { type: array, items: { type: string }, description: 已配置键名（不含值） }
    SyncTargetInput:
      type: object
      required: [name, platform, config]
      properties:
        name: { type: string }
        platform: { type: string, enum: [github, vercel, cloudflare] }
        project_id: { type: string, description: 缺省 = 组织内全部项目 }
        env_slug: { type: string, default: '*' }
        config:
          type: object
          additionalProperties: { type: string }
          description: |
            github: token/owner/repo/environment?(repo 级留空)；
            vercel: token/project_id/team_id?/targets?(逗号分隔，默认 production)；
            cloudflare: token/account_id/script_name
    SyncRun:
      type: object
      properties:
        id: { type: string }
        target_id: { type: string }
        secret_key: { type: string }
        action: { type: string }
        fingerprint: { type: string, description: 推送内容 sha256 前 16 hex（不含明文） }
        status: { type: string, enum: [pending, success, failed, dead] }
        trigger_type: { type: string, enum: [event, manual] }
        attempts: { type: integer }
        error: { type: string }
        created_at: { type: string }
        updated_at: { type: string }
...